14/7/2025

+++ qlnv/src/main/java/com/example/qlnv/entity/ChamCong.java
@@ -21,7 +21,7 @@ public class ChamCong {
 
     private int thang;
     private int nam;
-    private int soNgayCong;
+    private double soNgayCong;
     private int soNgayNghi;
 
     private Double luongThucNhan;
